requirements.txt
streamlit
pytesseract
pdf2image
pandas
openpyxl
opencv-python-headless
Pillow
